# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Management API",
    "version": "1.0.0",
    "description": "API for managing employee tasks and work plans"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://workplan-api.onrender.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "object"
              }
            ]
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "tools": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "enum": [
              "HIGH",
              "MEDIUM",
              "LOW"
            ]
          },
          "time": {
            "type": "string",
            "format": "time"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "IN_PROGRESS",
              "COMPLETED"
            ]
          },
          "constraints": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/hos/plan/tasks": {
      "post": {
        "tags": [
          "HOS Actions"
        ],
        "summary": "Fetch Tasks",
        "description": "Get tasks with optional filters",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "week": {
                    "type": "string",
                    "enum": [
                      "WEEK_1",
                      "WEEK_2",
                      "WEEK_3",
                      "WEEK_4"
                    ]
                  },
                  "month": {
                    "type": "string",
                    "enum": [
                      "JANUARY",
                      "FEBRUARY",
                      "MARCH",
                      "APRIL",
                      "MAY",
                      "JUNE",
                      "JULY",
                      "AUGUST",
                      "SEPTEMBER",
                      "OCTOBER",
                      "NOVEMBER",
                      "DECEMBER"
                    ]
                  },
                  "priorityFilter": {
                    "type": "string",
                    "enum": [
                      "HIGH",
                      "MEDIUM",
                      "LOW"
                    ]
                  },
                  "statusFilter": {
                    "type": "string",
                    "enum": [
                      "PENDING",
                      "IN_PROGRESS",
                      "COMPLETED"
                    ]
                  }
                }
              },
              "example": {
                "week": "WEEK_2",
                "month": "AUGUST",
                "priorityFilter": "HIGH",
                "statusFilter": "PENDING"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tasks retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Task"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Staff Actions"
        ],
        "summary": "Update Task",
        "description": "Update task status and completion details",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "plan_id": {
                    "type": "integer"
                  },
                  "task_id": {
                    "type": "integer"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "PENDING",
                      "IN_PROGRESS",
                      "COMPLETED"
                    ]
                  },
                  "completed": {
                    "type": "string",
                    "enum": [
                      "YES",
                      "NO"
                    ]
                  },
                  "reason": {
                    "type": "string"
                  }
                },
                "required": [
                  "plan_id",
                  "task_id",
                  "status"
                ]
              },
              "example": {
                "plan_id": 2,
                "task_id": 16,
                "status": "COMPLETED",
                "completed": "NO",
                "reason": "lack of funds"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/hou/dashboard": {
      "parameters": [],
      "post": {
        "summary": "Dashboard",
        "parameters": [],
        "responses": {},
        "tags": [
          "HOU"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "role_filter": "",
                "status_filter": ""
              }
            }
          }
        }
      }
    },
...
    "/api/hou/dashboard/staff/report/approve": {
      "parameters": [],
      "post": {
        "summary": "Performance report approval",
        "parameters": [],
        "responses": {},
        "tags": [
          "HOU"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "staff_id": {
                    "type": "string",
                    "format": "utc-millisec"
                  },
                  "plan_id": {
                    "type": "string",
                    "format": "utc-millisec"
                  }
                }
              },
              "example": {
                "staff_id": "4",
                "plan_id": "10"
              }
            }
          }
        }
      }
    }
  }
}